% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{entity_list}
\alias{entity_list}
\title{Entity list in the space_objects dataset}
\usage{
entity_list()
}
\value{
Entity list
}
\description{
Print list of entities that is in space_objects the dataset
}
\examples{
entity_list()
}
