% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Launch a shiny app}
\usage{
run_app()
}
\value{
Return Shiny app
}
\description{
This calls the shiny app based on the space_object dataset with analysis shown inside.
}
\examples{
\dontrun{
run_app()
}

}
