% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space-object.R
\docType{data}
\name{space_objects}
\alias{space_objects}
\title{Space Objects}
\format{
A data frame with 1,175 rows and 4 variables.
\describe{
\item{entity}{Country name, continent, and other entity}
\item{code}{Entity code}
\item{year}{Year}
\item{num_objects}{Annual number of objects launched into space}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2024/2024-04-23/outer_space_objects.csv}
}
\usage{
space_objects
}
\description{
Data for yearly number of objects launched into outer space of every entity
}
\keyword{datasets}
