% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{year_entity}
\alias{year_entity}
\title{Year range of selected entity}
\usage{
year_entity(entity)
}
\arguments{
\item{entity}{The entity name obtained from \code{entity_list()}}
}
\value{
Return range of the year of the selected entity in the space_objects dataset
}
\description{
Print the range of the year for selected entity
}
\examples{
year_entity("United States")
}
