% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceobs-package.R
\docType{package}
\name{spaceobs-package}
\alias{spaceobs}
\alias{spaceobs-package}
\title{spaceobs Package}
\description{
This package contain a yearly number of objects launched into outer space dataset
with shiny dashboard app showing the analysis of this dataset.
Learn more about \code{spaceobs} from the vignette: \code{browseVignettes(package = "spaceobs")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://etc5523-2024.github.io/assignment-4-packages-and-shiny-apps-suppawish11/}
}

}
\author{
\strong{Maintainer}: Suppawish Siriamnat \email{ssir0015@student.monash.edu}

}
