% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{year_entity}
\alias{year_entity}
\title{Year range of selected entity}
\usage{
year_entity(entity)
}
\arguments{
\item{entity}{}
}
\value{
Return range of the year of the selected entity in the space_objects dataset
}
\description{
Year range of selected entity
}
\examples{
year_entity()
}
